move workflow model to a submodule
