+++ src/github/index.ts
@@ -6,3 +6,5 @@ export * from './mergify';
 export * from './pr-template';
 export * from './mergify';
 export * from './auto-merge';
+
+export * as workflows from './workflows-model';
